Create Easy Version ( Count triplets Equal Xor )
Link: https://leetcode.com/problems/count-triplets-that-can-form-two-arrays-of-equal-xor/submissions/
/*
a == b means, that a ^ b == 0. Using this we can search for all contiguous arrays, with "0" xor of all it's elements. After finding the one we can add "size of this array - 1" to result variable because it is the number of ways we can split this arrays to "a" and "b".
*/
a = XOR( [i ...... k] ) 
b = XOR( [k+1 ......j] )
If a == b then XOR( [ i ...... j ] ) = 0
No of triplets = No of ways to split the above array = j-i ( length - 1)